refactor(settlements): split settlementreadinesslistpanel
Split the readiness list into focused components/modules, such as query panel, table, row, readiness
badge, manual readiness control, auto-ready control, and display text helpers.

fix #121
